Don't log events during unit tests
EventLogging should be disabled during all of our unit tests but, for
now, stub Schema.prototype.log during the CollectionsContentOverlay
test, which was logging a GatherClicks event.

This is a workaround for a bug in the Gather mediawiki-extensions-qunit
build wherein the GatherClicks schema isn't being registered, i.e. the
EventLogging::logEvent method isn't callable, but the EventLogging
extension is definitely loaded, which is obvious from the error itself.

Bug: T106759
Change-Id: Ic53bb343a9b703003a0310b74d8a9c5571defd13